# swagger-output.json
{"info":{"title":"Simple rest api for todo app","version":"1.0.0","contact":{"name":"Vitalya","email":"vitalya18@icloud.com"}},"swagger":"2.0","paths":{"/task":{"post":{"description":"Create new task","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"Task","schema":{"$ref":"#/definitions/Task"}}],"responses":{"200":{"description":"It's ok and task created"},"400":{"description":"It's bad request"},"500":{"description":"It's internal server error"}}},"get":{"description":"Get all task","consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"Return all task","schema":{"type":"array","items":{"$ref":"#/definitions/Task"}}},"400":{"description":"It's bad request"},"500":{"description":"It's internal server error"}}}},"/task/{id}":{"get":{"description":"Get task by id","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"path","type":"string","name":"id","description":"Id of task","required":"true,"}],"responses":{"200":{"description":"Return task by id","schema":{"$ref":"#/definitions/Task"}},"400":{"description":"It's bad request"},"500":{"description":"It's internal server error"}}},"post":{"description":"Edit existing task","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"path","type":"string","name":"id","description":"Id of task","required":"true,"},{"in":"body","name":"Task","schema":{"$ref":"#/definitions/Task"}}],"responses":{"200":{"description":"Return updated task","schema":{"$ref":"#/definitions/Task"}},"400":{"description":"It's bad request"},"500":{"description":"It's internal server error"}}},"delete":{"description":"Remove task by id","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"path","type":"string","name":"id","description":"Id of task","required":"true,"}],"responses":{"200":{"description":"Success removed"},"400":{"description":"It's bad request"},"500":{"description":"It's internal server error"}}}}},"definitions":{"Task":{"type":"object","required":["content"],"properties":{"content":{"type":"string","example":"Walk with dog"},"date":{"type":"Date","default":"Current date","example":"2021-08-04T11:36:40.235+00:00","info":"Not send, always set default"}}}},"responses":{},"parameters":{},"securityDefinitions":{},"tags":[]}